Refactoring options scripts
