+++ src/common/settings.js
@@ -1,0 +1,16 @@
+export const defaults = {
+    "IN PROGRESS": {
+        backgroundColour: "#F0B719",
+        textColour: "#000000"
+    },
+    "BACKLOG": {
+        backgroundColour: "#ff4747",
+        textColour: "#000000"
+    },
+    "DONE": {
+        backgroundColour: "#168500",
+        textColour: "#FFFFFF"
+    }
+};
+
+export const rowCount = 0;
